Change default package selection for simpleton case
It used to be that Text::Template would look one level up the stack.

Now that Text::Template::fill_in calls
Text::Template::Filler::fill_in, one level is wrong.  Two levels would
also be wrong if Filler::fill_in were called directly from user code.
So we have a loop that scans up the stack looking for the first
non-Text::Template module.  That should cover the common cases; if it
fails, there are still plenty of escape hatches.

But this code ought to be moved into a method.